+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## [v1.4.0](https://github.com/amirHossein5/dropy/compare/v1.3.0...v1.4.0)
+
+### Added
+
+-   `syncTogglers` option.
+
+
 ## [v1.3.0](https://github.com/amirHossein5/dropy/compare/v1.2.0...v1.3.0)
 
 ### Added